Added a function called `enableRawMode()`
This is useful in canonical mode, but really gets in the way when we are
trying to carefully render a user interface in raw mode. So we turn it
off. This program does the same thing as the one in the previous step,
it just doesn’t print what you are typing. You may be familiar with this
mode if you’ve ever had to type a password at the terminal, when using
sudo for example.