Disable `Ctrl-V`
On some systems, when you type Ctrl-V, the terminal
waits for you to type another character and then sends
that character literally. For example, before we disabled
Ctrl-C, you might’ve been able to type Ctrl-V and then Ctrl-C to
input a 3 byte. We can turn off this feature using the IEXTEN flag.