Fix `Ctrl-M`
If you run the program now and go through the whole alphabet while holding down
Ctrl, you should see that we have every letter except M. Ctrl-M is weird: it’s
being read as 10, when we expect it to be read as 13, since it is the 13th
letter of the alphabet, and Ctrl-J already produces a 10. What else produces 10?
The Enter key does.