A timeout for `read()`
Currently, read() will wait indefinitely for input from the keyboard before it
returns. What if we want to do something like animate something on the screen
while waiting for user input? We can set a timeout, so that read() returns if it
doesn’t get any input for a certain amount of time.